% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvb.R
\name{cross_validate}
\alias{cross_validate}
\title{Cross-validation for fitted models}
\usage{
cross_validate(
  folds,
  sigma,
  rho,
  X,
  y,
  add.intercept = TRUE,
  reps = 50,
  seed = 0,
  MAXITER = 100,
  EPS = 1e-05,
  verboseIter = FALSE
)
}
\arguments{
\item{sigma}{A hyperparameter tuning the regression coefficients' Normal(0,\code{sigma^2}) prior density.}

\item{rho}{A hyperparameter representing the sparsity of the Bernoulli(\code{rho}) prior density for the model selection mask.}

\item{X}{A matrix or vector containing the variables in the model.}

\item{y}{A vector containing the observed binary response for fitting the model.}

\item{add.intercept}{An optional logical for whether to add an intercept column of \code{1}s to the \code{X} input.}

\item{reps}{The number of repetitions for cross-validation.}

\item{seed}{The seed for drawing the folds from the supplied data.}

\item{MAXITER}{The maximum number of iterations allowed before concluding that the estimated coefficients do not converge.}

\item{EPS}{The threshold for residuals under which the model coefficients are considered converged.}

\item{verboseIter}{An optional logical for whether to print information about each cross-validation test.}
}
\value{
A vector of cross-validation errors corresponding to each repetition.
}
\description{
Estimates the MSE of the model via repeated cross-validation.
}
\examples{
# Initial Data
X = c(66,70,69,68,67,72,73,70,57,63,70,78,67)
y = c(0,1,0,0,0,0,0,0,1,1,1,0,0)

# Perform cross-validation
cross_validate(folds = 5, sigma = 10, rho = 0.1, X = X, y = y,
reps = 3, seed = 0, add.intercept = TRUE, MAXITER = 10, EPS = 1e-5, verboseIter = TRUE)
}
