% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvb.R
\name{predict_cvb}
\alias{predict_cvb}
\title{Classifier for fitted models}
\usage{
predict_cvb(fitted_model, X, add.intercept = NA)
}
\arguments{
\item{fitted_model}{A fitted model from \code{cvb_model}}

\item{add.intercept}{An optional logical for whether to add an intercept column of 1s to the X input.
If left blank, the function will assume the option used to train the model.}

\item{data}{A matrix or vector containing new observations for classification.}

\item{MAXITER}{The maximum number of iterations allowed before concluding that the estimated coefficients do not converge.}

\item{EPS}{The threshold for residuals under which the model coefficients are considered converged.}
}
\value{
A vector of predictions.
}
\description{
Uses a fitted logistic model from \code{cvb_model} and new data to perform binary classification. The probability threshold is \eqn{0.5}.
}
\examples{
# Initial Data
X = c(66,70,69,68,67,72,73,70,57,63,70,78,67)
y = c(0,1,0,0,0,0,0,0,1,1,1,0,0)
new.X = c(66,70,68,59)

# Fit the model
cvbdl::cvb_model(sigma = 10, rho = 0.1, X, y,
                add.intercept = TRUE, MAXITER = 100, EPS = 1e-5, verboseIter = FALSE) -> m; m

# Perform Prediction
cvbdl::predict_cvb(fitted_model = m, X = new.X)

}
